Increase timeout to (hopefully) avoid deadlock
There's a deadlock in starting and stopping rtspsrc. Stopping rtspsrc blocks
waiting for the internal task to complete while holding the state lock. If the
task is still executing _play when stop is called, _play will try to grab the
state lock too, which deadlocks as it's already taken by _stop. Grmbl.